requirements.txt
websockets==11.0.3
pyaudio==0.2.13
janus==0.7.0
streamlit==1.24.1
streamlit_autorefresh==1.0.1
python-dotenv==1.0.0
